Rename package lib to imageconv
