+++ kadai1/misonog/imageconv/files.go
@@ -1,4 +1,4 @@
-package lib
+package imageconv
 
 import (
 	"image"
