+++ kadai1/misonog/imageconv/files_test.go
@@ -1,4 +1,4 @@
-package lib
+package imageconv
 
 import (
 	"reflect"
