+++ kadai1/misonog/imageconv/imgconv.go
@@ -1,4 +1,4 @@
-package lib
+package imageconv
 
 import (
 	"fmt"
