+++ kadai1/misonog/imageconv/imgconv_test.go
@@ -1,4 +1,4 @@
-package lib
+package imageconv
 
 import "testing"
